\name{bargaining}
\docType{data}
\alias{bargaining}
\title{Amount sent in the Ultimatum Game}
\description{
  The Ultimatum game was played separately in four different countries. This data contains
the offers of 30 students in Israel and 27 in the United States on a scale from 0 to 10.
This dataset is taken from Roth et al. (1991). 
}
\usage{bargaining}
\format{A data frame containing 30 observations for Israel and 27 for the US.}
\references{Roth, A. E., Prasnikar, V., Okuno-Fujiwara, M., & Zamir,
  S. (1991). Bargaining and market behavior in Jerusalem, Ljubljana,
  Pittsburgh, and Tokyo: An experimental study. The American Economic
  Review, 1068-1095.
  }
\keyword{datasets}
