\name{french}
\docType{data}
\alias{french}
\title{Minority share holder scores of countries with civil law
  with/without french origin}
\description{
  This data gives minority share holder scores in countries with civil
  law, but differentiating countries with/without french origin.
}
\usage{french}
\format{A dataframe containing 32 observations of countries with french
  origin and 19 countries without french origin.}
\source{Karl Schlag}
\references{Karl Schlag}
\keyword{datasets}
