\name{mshscores}
\docType{data}
\alias{mshscores}
\title{Minority share holder scores for common and civil law countries}
\description{
  This data gives the minority share holder scores in countries with
  civil or common law.
}
\usage{mshscores}
\format{A dataframe containing 51 observations for civil law and 21 for
  common law.}
\source{Karl Schlag}
\references{Karl Schlag}
\keyword{datasets}
