\name{uncertainty}
\docType{data}
\alias{uncertainty}
\title{Uncertainty in a game theoretical experiment.}
\description{
  In an experiment, subjects played a similar game twice. Choices could
  be between 110 and 170. Each time, before they made their own choice,
  they had to indicate an interval [L, U] that they believed would
  contain the choice of their opponent. They paid some additional monefy
  if the choice of their opponent was in the interval they specified,
  and were paid more the smaller this interval was. So the width W_i of
  this interval in round i gives an indication of how uncertain they are
  in round i. The data contains the interval width in round 1 and 2.
}
\usage{uncertainty}
\format{A dataframe containing the 25 intervals in each round of the game.}
\source{Karl Schlag}
\references{Karl Schlag}
\keyword{datasets}
