\name{bargaining}
\docType{data}
\alias{bargaining}
\title{Shared money amount in an Ultimatum game}
\description{
  The Ultimatum game was played with 30 individuals in Israel and the
  United States. This dataset gives the amount of money the first
  person offered to the second person. 
}
\usage{bargaining}
\format{A data frame containing 30 observations for Israel and the US.}
\source{Karl Schlag}
\references{Karl Schlag}
\keyword{datasets}
