\name{bargaining}
\docType{data}
\alias{bargaining}
\title{Amount sent in the Ultimatum Game}
\description{
  The Ultimatum game was played separately in four different countries. This data contains
the offers of 30 students in Israel and 27 in the United States on a scale from 0 to 10.
This dataset is taken from Roth et al. (1991). 
}
\usage{bargaining}
\format{A data frame containing 30 observations for Israel and 27 for the US.}
\references{Roth et al. (1991), American Economic Review, Vol. 81, No. 5 (Dec., 1991), pp. 1068-1095.}
\keyword{datasets}
