% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npExact-package.R
\docType{data}
\name{bargaining}
\alias{bargaining}
\title{Amount sent in the Ultimatum Game}
\format{A data frame containing 30 observations for Israel and 27 for the
US.}
\description{
The Ultimatum game was played separately in four different countries. This
data contains the offers of 30 students in Israel and 27 in the United
States on a scale from 0 to 10. This dataset is taken from Roth et al.
(1991).
}
\references{
Roth, A. E., Prasnikar, V., Okuno-Fujiwara, M., & Zamir, S.
(1991). Bargaining and market behavior in Jerusalem, Ljubljana, Pittsburgh,
and Tokyo: An experimental study. The American Economic Review, 1068-1095.
}
\keyword{datasets}

