% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npExact-package.R
\docType{data}
\name{french}
\alias{french}
\title{Indices of minority shareholder protection of countries with civil law with
and without french origin.}
\format{A dataframe containing 32 observations of countries with french
origin and 19 countries without french origin.}
\description{
This data contains the indices of minority shareholder protection on a
scale from 0 to 1 in 51 countries with civil law, differentiating between
those with (32 observations) and those without (19 observations) french
origin. A higher value of the index means that country is more protected.
The data set is taken from Djankov et al. (2008).
}
\references{
Djankov, S., La Porta, R., Lopez-de-Silanes, F., & Shleifer, A.
(2008). The law and economics of self-dealing. Journal of financial
economics, 88(3), 430-465.
}
\keyword{datasets}

