% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npExact-package.R
\docType{data}
\name{mshscores}
\alias{mshscores}
\title{Indices of minority shareholder protection of countries with common and
with civil law.}
\format{A dataframe containing 51 observations for civil law and 21 for
common law.}
\description{
This data contains the indices of minority shareholder protection on a
scale from 0 to 1 in 51 countries with civil law and 21 countries with
common loaw. A higher value of the index means that country is more
protected. The data set is taken from Djankov et al. (2008).
}
\references{
Djankov, S., La Porta, R., Lopez-de-Silanes, F., & Shleifer, A.
(2008). The law and economics of self-dealing. Journal of financial
economics, 88(3), 430-465.
}
\keyword{datasets}

