\name{npExact-package}
\alias{npExact-package}
\alias{npExact}
\docType{package}
\title{Nonparametric hypothesis tests}
\description{
\code{npExact} provides hypothesis tests, which do not make assumptions
on the underlying distributions.
}
\details{
\tabular{ll}{
Package: \tab npExact\cr
Type: \tab Package\cr
Version: \tab 0.2 \cr
Date: \tab 2013-09-26 \cr
License: \tab GPL-2\cr
}
xxx Please include some information about the most important functions of
the package here
}
\author{
Karl Schlag, Oliver Reiter, Peter Saffert, Christian Pechhacker, Simona
Jokubauskaite, Tautvilas Janusauskas
}
\references{
  Karl Schlag (2012).
  Exact Hypothesis Testing without Assumptions -
  New and Old Results not only for Experimental Game Theory. Available
  at \url{http://homepage.univie.ac.at/karl.schlag/research/statistics/exacthypothesistesting8.pdf}
}
%% \keyword{nonparametric, distribution-free}

\seealso{
  \url{http://homepage.univie.ac.at/karl.schlag/research/statistics.html}
}
% \examples{
% }
