\name{npExact-package}
\alias{npExact-package}
\alias{npExact}
\docType{package}
\title{Nonparametric hypothesis tests}
\description{
\code{npExact} provides hypothesis tests, which do not make assumptions
on the underlying distributions.
}
\details{
\tabular{ll}{
Package: \tab npExact\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-17\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Karl Schlag

Maintainer: Oliver Reiter
}
\references{
  Karl Schlag (2012).
  Exact Hypothesis Testing without Assumptions -
  New and Old Results not only for Experimental Game Theory
}
\keyword{nonparametric}
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
\examples{
npCorr(runif(20), runif(20))
}
