\name{pain}
\docType{data}
\alias{pain}
\title{Pain experienced before and after a knie operation}
\description{
  There are two ways to determine where to start an operation on a knee,
  either with a computer or manually. The data describes the pain
  experienced by the patients before and after the surgery.
}
\usage{pain}
\format{A dataframe containing 50 observations. Column "pc" indicates if
a computer was used (coded with "1") or not (coded with "0")}
\references{Sabeti-Aschraf, M., Dorotka, R., Goll, A., & Trieb, K. (2005). Extracorporeal shock wave therapy in the treatment of calcific tendinitis of the rotator cuff. The American journal of sports medicine, 33(9), 1365-1368.}
\keyword{datasets}
