% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npExact-package.R
\docType{data}
\name{pain}
\alias{pain}
\title{Pain experienced before and after a knie operation}
\format{A dataframe containing 50 observations. Column "pc" indicates if a
computer was used (coded with "1") or not (coded with "0")}
\description{
There are two ways to determine where to start an operation on a knee,
either with a computer or manually. The data describes the pain experienced
by the patients before and after the surgery.
}
\references{
Sabeti-Aschraf, M., Dorotka, R., Goll, A., & Trieb, K. (2005).
Extracorporeal shock wave therapy in the treatment of calcific tendinitis
of the rotator cuff. The American journal of sports medicine, 33(9),
1365-1368.
}
\keyword{datasets}

